+++ Source/OxyPlot/Annotations/PathAnnotation.cs
@@ -103,6 +103,30 @@ namespace OxyPlot.Annotations
         public double TextPadding { get; set; }
 
         /// <summary>
+        /// Gets or sets the padding of the border rectangle.
+        /// </summary>
+        /// <value>The border padding.</value>
+        public OxyThickness BorderPadding { get; set; }
+
+        /// <summary>
+        /// Gets or sets the fill color of the border rectangle.
+        /// </summary>
+        /// <value>The border background color.</value>
+        public OxyColor BorderBackground { get; set; }
+
+        /// <summary>
+        /// Gets or sets the stroke color of the border rectangle.
+        /// </summary>
+        /// <value>The border stroke color.</value>
+        public OxyColor BorderStroke { get; set; }
+
+        /// <summary>
+        /// Gets or sets the stroke thickness of the border rectangle.
+        /// </summary>
+        /// <value>The border stroke thickness.</value>
+        public double BorderStrokeThickness { get; set; }
+
+        /// <summary>
         /// Gets or sets the text orientation.
         /// </summary>
         /// <value>The text orientation.</value>
@@ -252,6 +276,20 @@ namespace OxyPlot.Annotations
                     if (this.TextPosition.IsDefined())
                     {
                         angle = this.TextRotation;
+                    }
+
+                    var textSize = rc.MeasureText(this.Text, this.ActualFont, this.ActualFontSize, this.ActualFontWeight);
+
+                    var textBounds = TextAnnotation.GetTextBounds(position, textSize, this.BorderPadding, angle, ha, va);
+
+                    if ((angle % 90).Equals(0))
+                    {
+                        var actualRect = new OxyRect(textBounds[0], textBounds[2]);
+                        rc.DrawRectangle(actualRect, this.BorderBackground, this.BorderStroke, this.BorderStrokeThickness, this.EdgeRenderingMode);
+                    }
+                    else
+                    {
+                        rc.DrawPolygon(textBounds, this.BorderBackground, this.BorderStroke, this.BorderStrokeThickness, this.EdgeRenderingMode);
                     }
 
                     rc.DrawText(